requirements.txt
fastapi
uvicorn
python-multipart
rembg
pillow
numpy
onnxruntime
supabase
pydantic
python-dotenv